requirements.txt
scikit-learn==1.1.0
seaborn==0.12.2
pandas==2.0.3
Pillow==10.0.1
numpy==1.24.4
torch==2.1.0